refactor(kernels): split ffi.rs into per-model submodules
Break the 1376-line monolithic ffi.rs into a 12-line dispatcher plus
ffi/{shared,qwen35,deepseek}.rs (33/14/60 symbols), grouped by the model
that owns each kernel. Public paths (pegainfer_kernels::ffi::<symbol>) and
the core->model re-export chain are unchanged, so there are zero call-site
edits. Pure declaration move; no kernel sources, build steps, or signatures
touched.

Addresses issue #127: FFI bindings are now predictable to locate and far
less merge-conflict prone.
