+++ pegainfer-kernels/src/ffi/qwen35.rs
@@ -1,0 +1,230 @@
+use super::Half;
+use cudarc::driver::sys::{CUresult, CUstream};
+
+// Qwen3.5-4B private kernels (hybrid linear + HD256 full attention).
+// Sources: csrc/{gated_delta_rule,conv1d,prefill_attention_hd256,paged_attention}.cu
+unsafe extern "C" {
+    // Qwen3.5 full-attention prefill prep: Q/K norm + partial RoPE + KV cache write.
+    pub fn prefill_attention_hd256_prep_cuda(
+        q_full_batch: *const Half,
+        k_batch: *const Half,
+        v_batch: *const Half,
+        q_norm_weight: *const Half,
+        k_norm_weight: *const Half,
+        cos_cache: *const Half,
+        sin_cache: *const Half,
+        q_batch_out: *mut Half,
+        k_cache: *mut Half,
+        v_cache: *mut Half,
+        num_q_heads: i32,
+        num_kv_heads: i32,
+        seq_len: i32,
+        start_pos_ptr: *const i32,
+        rotary_dim: i32,
+        rms_eps: f32,
+        max_seq_len: i32,
+        stream: CUstream,
+    );
+
+    // Apply sigmoid(gate) from interleaved q_full onto attention output in-place.
+    pub fn attention_gate_batch_hd256_cuda(
+        q_full_batch: *const Half,
+        attn_out: *mut Half,
+        num_q_heads: i32,
+        seq_len: i32,
+        stream: CUstream,
+    );
+
+    pub fn qk_norm_partial_rope_batched_decode_hd256_cuda(
+        q_full_batch: *const Half,
+        k_batch: *mut Half,
+        q_norm_weight: *const Half,
+        k_norm_weight: *const Half,
+        cos_cache: *const Half,
+        sin_cache: *const Half,
+        positions: *const i32,
+        q_batch_out: *mut Half,
+        num_q_heads: i32,
+        num_kv_heads: i32,
+        batch_size: i32,
+        rotary_dim: i32,
+        rms_eps: f32,
+        stream: CUstream,
+    );
+
+    // Gated delta rule recurrent decode (single step)
+    pub fn gated_delta_rule_decode_cuda(
+        qkv: *const Half,
+        b_proj: *const Half,
+        a_proj: *const Half,
+        dt_bias: *const Half,
+        A_log: *const f32,
+        state: *mut f32,
+        output: *mut Half,
+        num_key_heads: i32,
+        num_value_heads: i32,
+        key_dim: i32,
+        val_dim: i32,
+        stream: CUstream,
+    );
+
+    // Causal depthwise conv1d prefill (parallel over sequence)
+    pub fn conv1d_prefill_cuda(
+        x_seq: *const Half,
+        conv_weight: *const Half,
+        conv_state: *mut Half,
+        out_seq: *mut Half,
+        num_channels: i32,
+        seq_len: i32,
+        kernel_size: i32,
+        stream: CUstream,
+    );
+
+    pub fn gated_delta_rule_prefill_chunk_prepare_cuda(
+        qkv: *const Half,
+        b_proj: *const Half,
+        a_proj: *const Half,
+        dt_bias: *const Half,
+        a_log: *const f32,
+        q_out: *mut Half,
+        k_out: *mut Half,
+        v_out: *mut Half,
+        g_out: *mut f32,
+        beta_out: *mut f32,
+        num_key_heads: i32,
+        num_value_heads: i32,
+        qkv_dim: i32,
+        seq_len: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    pub fn gated_delta_rule_prefill_chunk_cumsum_cuda(
+        g_in: *const f32,
+        g_out: *mut f32,
+        seq_len: i32,
+        num_value_heads: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    pub fn gated_delta_rule_prefill_chunk_a_cuda(
+        k: *const Half,
+        g_cumsum: *const f32,
+        beta: *const f32,
+        a_tril: *mut f32,
+        seq_len: i32,
+        num_value_heads: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    pub fn gated_delta_rule_prefill_chunk_solve_cuda(
+        a_tril: *const f32,
+        a_inv: *mut Half,
+        seq_len: i32,
+        num_value_heads: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    pub fn gated_delta_rule_prefill_chunk_recompute_cuda(
+        k: *const Half,
+        v: *const Half,
+        beta: *const f32,
+        w: *mut Half,
+        u: *mut Half,
+        a_inv: *const Half,
+        g_cumsum: *const f32,
+        seq_len: i32,
+        num_value_heads: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    // Chunk-wise GDR prefill stage 1 (Triton AOT): recurrent chunk-state update.
+    // Expected future inputs:
+    //   k / w: [seq_len, num_value_heads, 128] bf16
+    //   u / v_new: [seq_len, num_value_heads, 128] bf16
+    //   g_cumsum: [seq_len, num_value_heads] fp32
+    //   initial_state / final_state: [num_value_heads, 128, 128] fp32 in [H, K, V] (V contiguous)
+    //   chunk_state: [num_chunks, num_value_heads, 128, 128] fp32
+    pub fn gated_delta_rule_prefill_chunk_state_cuda(
+        k: *const Half,
+        w: *const Half,
+        u: *const Half,
+        g_cumsum: *const f32,
+        initial_state: *const f32,
+        chunk_state: *mut f32,
+        v_new: *mut Half,
+        final_state: *mut f32,
+        seq_len: i32,
+        num_value_heads: i32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    // Chunk-wise GDR prefill stage 2 (Triton AOT): chunk output accumulation.
+    // Expected future inputs:
+    //   q / k / v_new: [seq_len, num_value_heads, 128] bf16
+    //   chunk_state: [num_chunks, num_value_heads, 128, 128] fp32
+    //   g_cumsum: [seq_len, num_value_heads] fp32
+    //   output: [seq_len, num_value_heads * 128] bf16
+    pub fn gated_delta_rule_prefill_chunk_o_cuda(
+        q: *const Half,
+        k: *const Half,
+        v_new: *const Half,
+        chunk_state: *const f32,
+        g_cumsum: *const f32,
+        output: *mut Half,
+        seq_len: i32,
+        num_value_heads: i32,
+        scale: f32,
+        stream: CUstream,
+    ) -> CUresult;
+
+    // Paged attention decode for HEAD_DIM=256 (Qwen3.5-4B full-attention layers).
+    pub fn paged_attention_decode_cuda_hd256(
+        q: *const Half,
+        output: *mut Half,
+        kv_data: *const Half,
+        k_offset_elems: i64,
+        v_offset_elems: i64,
+        page_indices: *const i32,
+        page_indptr: *const i32,
+        last_page_len_d: *const i32,
+        request_indices: *const i32,
+        kv_tile_indices: *const i32,
+        kv_chunk_size_ptr: *const i32,
+        num_qo_heads: i32,
+        num_kv_heads: i32,
+        head_dim: i32,
+        page_size: i32,
+        batch_size: i32,
+        stride_page: i64,
+        sm_scale: f32,
+        stream: CUstream,
+    ) -> i32;
+
+    // Batch prefill with paged KV for HEAD_DIM=256 (Qwen3.5-4B multi-token prefill).
+    pub fn batch_prefill_paged_cuda_hd256(
+        q: *const Half,
+        output: *mut Half,
+        kv_data: *const Half,
+        k_offset_elems: i64,
+        v_offset_elems: i64,
+        page_indices: *const i32,
+        page_indptr: *const i32,
+        last_page_len_d: *const i32,
+        q_indptr: *const i32,
+        request_indices: *const i32,
+        qo_tile_indices: *const i32,
+        kv_tile_indices: *const i32,
+        kv_chunk_size_ptr: *const i32,
+        total_num_rows: *const u32,
+        num_qo_heads: i32,
+        num_kv_heads: i32,
+        head_dim: i32,
+        page_size: i32,
+        seq_len: i32,
+        batch_size: i32,
+        padded_batch_size: i32,
+        stride_page: i64,
+        sm_scale: f32,
+        stream: CUstream,
+    ) -> i32;
+}
